move test case from DataFrameReaderWriterSuite to JsonSuite
