+++ sql/core/src/test/scala/org/apache/spark/sql/execution/datasources/json/JsonSuite.scala
@@ -17,7 +17,7 @@
 
 package org.apache.spark.sql.execution.datasources.json
 
-import java.io.{File, StringWriter}
+import java.io.{File, FileOutputStream, OutputStreamWriter, StringWriter}
 import java.nio.charset.StandardCharsets
 import java.sql.{Date, Timestamp}
 import java.util.Locale
@@ -2033,5 +2033,28 @@ class JsonSuite extends QueryTest with SharedSQLContext with TestJsonData {
         assert(e.getMessage.contains("Cannot parse"))
       }
     }
+  }
+
+  test("SPARK-21610: Corrupt records are not handled properly when creating a dataframe " +
+    "from a file") {
+    val tempDir = Utils.createTempDir()
+    val file = new File(tempDir, "sample.json")
+    val writer = new OutputStreamWriter(new FileOutputStream(file), StandardCharsets.UTF_8)
+    val data =
+      """{"field": 1}
+        |{"field": 2}
+        |{"field": "3"}""".stripMargin
+    writer.write(data)
+    writer.close()
+
+    val schema = new StructType()
+      .add("field", ByteType)
+      .add("_corrupt_record", StringType)
+
+    val dfFromFile = spark.read.schema(schema).json(file.getAbsolutePath)
+    assert(dfFromFile.filter($"_corrupt_record".isNotNull).count() == 1)
+    assert(dfFromFile.filter($"_corrupt_record".isNull).count() == 2)
+
+    Utils.deleteRecursively(tempDir)
   }
 }
