+++ sql/core/src/test/scala/org/apache/spark/sql/test/DataFrameReaderWriterSuite.scala
@@ -17,8 +17,7 @@
 
 package org.apache.spark.sql.test
 
-import java.io.{File, FileOutputStream, OutputStreamWriter}
-import java.nio.charset.StandardCharsets
+import java.io.File
 import java.util.Locale
 import java.util.concurrent.ConcurrentLinkedQueue
 
@@ -775,28 +774,5 @@ class DataFrameReaderWriterSuite extends QueryTest with SharedSQLContext with Be
         }
       }
     }
-  }
-
-  test("SPARK-21610: Corrupt records are not handled properly when creating a dataframe " +
-    "from a file") {
-    val tempDir = Utils.createTempDir()
-    val file = new File(tempDir, "sample.json")
-    val writer = new OutputStreamWriter(new FileOutputStream(file), StandardCharsets.UTF_8)
-    val data =
-      """{"field": 1}
-        |{"field": 2}
-        |{"field": "3"}""".stripMargin
-    writer.write(data)
-    writer.close()
-
-    val schema = new StructType()
-      .add("field", ByteType)
-      .add("_corrupt_record", StringType)
-
-    val dfFromFile = spark.read.schema(schema).json(file.getAbsolutePath)
-    assert(dfFromFile.filter($"_corrupt_record".isNotNull).count() == 1)
-    assert(dfFromFile.filter($"_corrupt_record".isNull).count() == 2)
-
-    Utils.deleteRecursively(tempDir)
   }
 }
